logservice: various fixes
